datatype moved from decimal to double

+++ Project0/RestaurantService/IService.cs
@@ -16,9 +16,9 @@ namespace RestaurantService
 
     public interface IRestaurant
     {
-        void AddRating(decimal rating);
+        void AddRating(double rating);
         
-        void SearchByRating(decimal rating);
+        void SearchByRating(double rating);
         void SearchByName(string name);
         void PrintDetails(string resName);
